feat: add --sort flag to encode.py and refactor sorting logic
- Extracted sorting logic from decode.py into lib/sortlib.py for reusability.
- Added --sort flag to encode.py with support for name, color, type, and cmc.
- Updated README.md to document the new feature.
- Added unit tests for the shared sorting library.
